% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/amend_output.R
\name{amend_output}
\alias{amend_output}
\title{Format density covariate matrix}
\usage{
amend_output(
  TmbData,
  Report,
  Sdreport = NULL,
  year_labels = NULL,
  category_names = NULL
)
}
\description{
\code{amend_output} uses a formula interface to generate covariates
}
