% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Calc_Anisotropic_Mesh.R
\name{Calc_Anisotropic_Mesh}
\alias{Calc_Anisotropic_Mesh}
\title{Make mesh for distances among points}
\usage{
Calc_Anisotropic_Mesh(loc_x, loc_g, loc_i, Method, Extrapolation_List,
  anisotropic_mesh = NULL, fine_scale = FALSE, ...)
}
\arguments{
\item{loc_x}{location (eastings and northings in kilometers, UTM) for each sample or knot}

\item{Method}{spatial method determines ("Mesh" and "Grid" give}

\item{anisotropic_mesh}{OPTIONAL, anisotropic mesh (if missing, its recalculated from loc_x)}

\item{...}{Arguments passed to \code{INLA::inla.mesh.create}}
}
\value{
Tagged list containing distance metrics
}
\description{
\code{Calc_Anistropic_Mesh} builds a tagged list representing distances for isotropic or geometric anisotropic triangulated mesh
}
