% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_proportion.R
\name{calculate_proportion}
\alias{calculate_proportion}
\title{Calculate the compositional-expansion}
\usage{
calculate_proportion(TmbData, Index, Year_Set = NULL,
  Years2Include = NULL, strata_names = NULL, category_names = NULL,
  plot_legend = TRUE, DirName = paste0(getwd(), "/"),
  PlotName = "Proportion.png", interval_width = 1, width = 6,
  height = 6, ...)
}
\arguments{
\item{Index}{output from \code{FishStatsUtils::plot_biomass_index}}

\item{...}{list of settings to pass to \code{sdreport}}
}
\value{
Tagged list of output
\describe{
  \item{Prop_ctl}{Proportion of biomass for each category c, time t, and stratum l}
  \item{Neff_tl}{Effective sample size (median across categories) for each time t and stratum l}
}
}
\description{
\code{calculate_proportion} takes output from a VAST run and calculates the proportion of biomass in different categories
}
