% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_shapefile.R
\name{convert_shapefile}
\alias{convert_shapefile}
\title{Convert shapefile to extrapolation-grid}
\usage{
convert_shapefile(file_path, projargs = NULL, grid_dim_km = c(2, 2),
  make_plots = FALSE, quiet = TRUE)
}
\arguments{
\item{file_path}{path for shapefile on harddrive}

\item{projargs}{A character string of projection arguments; the arguments must be entered exactly as in the PROJ.4 documentation; if the projection is unknown, use \code{as.character(NA)}, it may be missing or an empty string of zero length and will then set to the missing value.}

\item{make_plots}{Boolean indicating whether to visualize inputs and outputs as maps}
}
\value{
extrapolation-grid
}
\description{
\code{convert_shapefile} reads in a shapefile and creates an extrapolation with a chosen resolution
}
\examples{
\dontrun{
 convert_shapefile( file_path="C:/Users/James.Thorson/Desktop/Work files/AFSC/2020-03 -- Add ICES grids/IBTS grids/BITS/Shapefile.shp", make_plots=TRUE )
}
}
