% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FishStatsUtils.R
\docType{package}
\name{FishStatsUtils}
\alias{FishStatsUtils}
\alias{FishStatsUtils-package}
\title{FishStatsUtils}
\description{
FishStatsUtils provides utilities (shared code and data) for spatio-temporal models (VAST, SpatialDeltaGLMM, ...)
}
\details{
Please see examples folder on github \url{http://github.com/james-thorson/VAST/tree/master/examples} for up-to-date examples
}

