% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_spatial_info.R
\name{make_spatial_info}
\alias{make_spatial_info}
\title{Build objects related to spatial information}
\usage{
make_spatial_info(n_x, Lon_i, Lat_i, LON_intensity = Lon_i,
  LAT_intensity = Lat_i, Extrapolation_List, Method = "Mesh",
  grid_size_km = 50, grid_size_LL = 1, fine_scale = FALSE,
  iter.max = 1000, randomseed = 1, nstart = 100,
  DirPath = paste0(getwd(), "/"), Save_Results = FALSE, ...)
}
\arguments{
\item{n_x, }{the number of vertices in the SPDE mesh (determines the spatial resolution when Method="Mesh")}

\item{Lon_i, }{Longitude for each sample}

\item{Lat_i, }{Latitude for each sample}

\item{LON_intensity, }{Longitude for each location to use during k-means algorithm to decide upon the location of knots}

\item{LAT_intensity, }{Latitude for each location to use during k-means algorithm to decide upon the location of knots}

\item{Extrapolation_List, }{the output from \code{Prepare_Extrapolation_Data_Fn}}

\item{Method, }{a character of either "Grid" or "Mesh" where "Grid" is a 2D AR1 process, and "Mesh" is the SPDE method with geometric anisotropy}

\item{grid_size_km, }{the distance between grid cells for the 2D AR1 grid (determines spatial resolution when Method="Grid") when not using \code{Method="Spherical_mesh"}}

\item{grid_size_LL, }{the distance between grid cells for the 2D AR1 grid (determines spatial resolution when Method="Grid") when using \code{Method="Spherical_mesh"}}

\item{iter.max}{the number of iterations used per k-means algorithm (default=1000)}

\item{randomseed}{a random number seed}

\item{nstart}{the number of times that the k-means algorithm is run while searching for the best solution (default=100)}

\item{DirPath}{a directory where the algorithm looks for a previously-saved output (default is working directory)}

\item{Save_Results}{a boolean stating whether to save the output (Default=TRUE)}

\item{..., }{additional arguments passed to \code{INLA::inla.mesh.create}}
}
\value{
Tagged list containing objects for running a VAST model
\describe{
  \item{MeshList}{A tagged list with inputs related to the SPDE mesh}
  \item{GridList}{A tagged list with inputs related to the 2D AR1 grid}
  \item{a_xl}{A data frame with areas for each knot and each strattum}
  \item{loc_UTM}{A data frame with the converted UTM coordinates for each sample}
  \item{Kmeans}{Output from \code{Calc_Kmeans} with knots for a triangulated mesh}
  \item{knot_i}{The knot associated with each sample}
  \item{Method}{The Method input (for archival purposes)}
  \item{loc_x}{The UTM location for each knot}
}
}
\description{
\code{make_spatial_info} builds a tagged list with all the spatial information needed for \code{Data_Fn}
}
