% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_model.R
\name{plot.fit_model}
\alias{plot.fit_model}
\title{Print parameter estimates}
\usage{
\method{print}{fit_model}(x, ...)
}
\arguments{
\item{x}{Output from \code{\link{fit_model}}}

\item{...}{Arguments passed to \code{\link{plot_results}}}
}
\description{
Print parameter estimates and standard errors.
}
