% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_biomass_index.R
\name{plot_biomass_index}
\alias{plot_biomass_index}
\title{Plot index of abundance}
\usage{
plot_biomass_index(
  TmbData,
  Sdreport,
  year_labels = NULL,
  years_to_plot = NULL,
  DirName = paste0(getwd(), "/"),
  PlotName = "Index",
  interval_width = 1,
  strata_names = NULL,
  category_names = NULL,
  use_biascorr = TRUE,
  plot_legend = TRUE,
  total_area_km2 = NULL,
  plot_log = FALSE,
  width = NULL,
  height = NULL,
  create_covariance_table = FALSE,
  Yrange = c(ifelse(plot_log == TRUE, NA, 0), NA),
  ...
)
}
\arguments{
\item{TmbData}{Formatted data inputs, from `VAST::Data_Fn(...)`}

\item{Sdreport}{Standard deviation outputs from TMB model via \code{sdreport(Obj)}}

\item{year_labels}{Year names for labeling panels}

\item{years_to_plot}{integer vector, specifying positions of \code{year_labels} for plotting (used to avoid plotting years with no data, etc.)}

\item{DirName}{Directory for saving plot and table}

\item{PlotName}{Name for plot}

\item{interval_width}{width for confidence intervals}

\item{strata_names}{names for spatial strata}

\item{category_names}{names for categories (if using package `VAST`)}

\item{use_biascorr}{Boolean, whether to use bias-corrected estimates if available}

\item{plot_legend}{Add legend for labelling colors}

\item{total_area_km2}{Total area for calculating a design-based estimator using one design-stratum (only recommended for model exploration)}

\item{plot_log}{Boolean, whether to plot y-axis in log-scale}

\item{width}{plot width in inches}

\item{height}{plot height in inches}

\item{...}{Other inputs to `par()`}
}
\value{
Return Tagged list of output
\describe{
  \item{Table}{table of index estimates by stratum and year, e.g., for including in an assessment model}
}
}
\description{
\code{plot_biomass_index} plots an index proportion to population abundance
}
\references{
For details regarding spatio-temporal index standardization see \url{https://doi.org/10.1093/icesjms/fsu243}

For details regarding fishing mortality and biomass reference points in models generating those, see \url{https://doi.org/10.1111/faf.12398}
}
