% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_clusters.R
\name{plot_clusters}
\alias{plot_clusters}
\title{Plot spatial clusters}
\usage{
plot_clusters(
  fit,
  map_list = NULL,
  var_name = "D_gct",
  transform_var = log,
  k = 4,
  method = "ward",
  working_dir = paste0(getwd(), "/"),
  format = "points",
  ...
)
}
\arguments{
\item{method}{Distance metric. Default \code{method="ward"} is very fast,
but can instead use \code{method="bcdist"} for small problems
\code{\link[ecodist]{bcdist}}}
}
\description{
\code{plot_clusters} plots the clusters on a map
}
\details{
Code to plot clusters to look for biogeographic shifts
}
