% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_index.R
\name{plot_index}
\alias{plot_index}
\title{Plot an index}
\usage{
plot_index(Index_ctl, sd_Index_ctl = array(0, dim(Index_ctl)),
  Year_Set = NULL, Years2Include = NULL, strata_names = NULL,
  category_names = NULL, scale = "uniform", plot_legend = NULL,
  DirName = paste0(getwd(), "/"), PlotName = "Index.png",
  interval_width = 1, width = NULL, height = NULL, xlab = "Year",
  ylab = "Index", bounds_type = "whiskers", col = NULL,
  col_bounds = NULL, Ymin = 0, ...)
}
\arguments{
\item{Index_ctl}{A matrix or array of time-series estimates for multiple categories \code{c}, years \code{t}, and strata \code{l}}

\item{sd_Index_ctl}{A matrix or array of variances for each estimate}

\item{Year_Set}{Year names for labeling panels}

\item{Years2Include}{integer vector, specifying positions of \code{Year_Set} for plotting (used to avoid plotting years with no data, etc.)}

\item{strata_names}{names for spatial strata}

\item{category_names}{names for categories (if using package `VAST`)}

\item{plot_legend}{Add legend for labelling colors}

\item{DirName}{Directory for saving plot and table}

\item{PlotName}{Name for plot}

\item{interval_width}{width for confidence intervals}

\item{width}{plot width in inches}

\item{height}{plot height in inches}

\item{Ymin}{lower bound for y-axis (use \code{Ymin=NA} for using the lower range of estimate intervals)}

\item{...}{list of settings to pass to \code{par} when making plot}
}
\description{
\code{plot_index} takes output from a VAST run and plots a panel figure of time-series estimates
}
