% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_maps.r
\name{plot_maps}
\alias{plot_maps}
\title{Plot standard maps}
\usage{
plot_maps(plot_set = 3, Report, PlotDF, Sdreport = NULL,
  TmbData = NULL, Panel = "Category", projargs = "+proj=longlat",
  MapSizeRatio = c(`Width(in)` = 4, `Height(in)` = 4), Res = 200,
  Legend = NULL, Year_Set = NULL, Years2Include = NULL,
  category_names = NULL, textmargin = NULL, quiet = FALSE,
  working_dir = paste0(getwd(), "/"), ...)
}
\arguments{
\item{plot_set}{integer-vector defining plots to create
\describe{
  \item{plot_set=1}{Probability of encounter/non-encounter}
  \item{plot_set=2}{Log-expected positive catch rate}
  \item{plot_set=3}{Log-predicted density (product of encounter probability and positive catch rates)}
  \item{plot_set=4}{Log-positive catch rates (rescaled)}
  \item{plot_set=5}{Log-predicted density (rescaled)}
  \item{plot_set=6}{Spatio-temporal variation in encounter probability}
  \item{plot_set=7}{Spatio-temporal variation in log-positive catch rates}
  \item{plot_set=8}{Linear predictor for encounter probability}
  \item{plot_set=9}{Linear predictor for positive catch rates}
  \item{plot_set=10}{Coefficient of variation for predicted density (available only if \code{Data_Fn(...,Options=c('SD_site_logdensity'=1,...))}}
  \item{plot_set=11}{Covariates that are included in the model}
  \item{plot_set=12}{Total biomass across all categories (only useful in a multivariate model)}
  \item{plot_set=13}{Covariate effects on encounter probability}
  \item{plot_set=14}{Covariate effects on positive catch rates}
}}

\item{Report}{tagged list of outputs from TMB model via \code{Obj$report()}}

\item{Sdreport}{Standard deviation outputs from TMB model via \code{sdreport(Obj)}}

\item{Panel}{Whether to plot years for a given category (\code{Panel="Category"}) or categories for a given year ((\code{Panel="Year"})  in each panel figure}

\item{MapSizeRatio}{Default size for each panel}

\item{Year_Set}{Year names for labeling panels}

\item{Years2Include}{integer vector, specifying positions of \code{Year_Set} for plotting (used to avoid plotting years with no data, etc.)}

\item{category_names}{character vector specifying names for different categories (only used for R package \code{VAST})}

\item{...}{arguments passed to \code{FishStatsUtils::plot_variable}}
}
\value{
Mat_xt a matrix (rows: modeled knots; column: modeled year) for plotted output of last element of \code{plot_set}
}
\description{
\code{plot_maps} plots a standard set of diagnostic maps
}
