% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_overdispersion.R
\name{plot_overdispersion}
\alias{plot_overdispersion}
\title{Plot estimated overdispersion}
\usage{
plot_overdispersion(
  filename1,
  filename2,
  Data,
  ParHat,
  Report,
  SD = NULL,
  Map = NULL,
  ControlList1 = list(Width = 8, Height = 4, Res = 200, Units = "in"),
  ControlList2 = ControlList1
)
}
\arguments{
\item{filename1}{filename (including absolute path) for overdispersion for component #1}

\item{filename2}{filename (including absolute path) for overdispersion for component #2}

\item{Data}{tagged list of input data}

\item{ParHat}{Tagged list of fitted data, e.g., from \code{ParHat <- obj$env$parList()}}

\item{Report}{output report, e.g., from \code{Report <- obj$report()}}

\item{SD}{standard deviation report, e.g., from \code{SD <- sdreport( obj )}}

\item{Map}{tagged list of map inputs}

\item{ControlList1}{plotting size for component #1}

\item{ControlIist2}{plotting size for component #2 (default: ControlList2 <- ControlList1)}
}
\value{
Tagged list containing inputs to function Build_TMB_Fn()
\describe{
  \item{Cov1_cc}{Covariance for component #1 (typically probability of occurence)}
  \item{Cov2_cc}{Covariance for component #2 (typically positive catch rates)}
}
}
\description{
\code{plot_overdispersion} plots sample and analytic estimates of overdispersion
}
