% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_residuals.R
\name{plot_residuals}
\alias{plot_residuals}
\title{Plot Pearson residuals on map}
\usage{
plot_residuals(Lat_i, Lon_i, TmbData, Report, Q,
  projargs = "+proj=longlat", working_dir = paste0(getwd(), "/"),
  spatial_list, extrapolation_list, Year_Set = NULL,
  Years2Include = NULL, zrange, ...)
}
\arguments{
\item{Lat_i}{Latitude for every observation \code{i}}

\item{Lon_i}{Longitude for every observation \code{i}}

\item{TmbData}{a tagged list of data inputs generated by \code{make_data}}

\item{Report}{tagged list of outputs from TMB model via \code{Obj$report()}}

\item{Q}{Output from \code{QQ_Fn}}

\item{projargs}{A character string of projection arguments; the arguments must be entered exactly as in the PROJ.4 documentation; if the projection is unknown, use \code{as.character(NA)}, it may be missing or an empty string of zero length and will then set to the missing value.}

\item{Year_Set}{Year names for labeling panels}

\item{Years2Include}{integer vector, specifying positions of \code{Year_Set} for plotting (used to avoid plotting years with no data, etc.)}

\item{zrange}{lower and upper bounds for plotted value of Pearson residuals, used to change plotting scale to eliminate influence of outliers on plotting scale}

\item{...}{arguments passed to \code{FisshStatsUtils::plot_variable}}

\item{savdir}{directory to use when saving results}
}
\value{
A tagged list of Pearson residuals
\describe{
  \item{Q1_xt}{Matrix of average residuals for encounter/non-encounter component by site \code{x} and year \code{t}}
  \item{Q2_xt}{Matrix of average residuals for positive-catch-rate component by site \code{x} and year \code{t}}
}
}
\description{
\code{plot_residuals} shows average Pearson residual for every knot for encounter probability and positive catch rate components
}
