% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_variable.R
\name{plot_variable}
\alias{plot_variable}
\title{Plot results on a multi-panel figure of maps}
\usage{
plot_variable(Y_gt, map_list, panel_labels, projargs = "+proj=longlat",
  map_resolution = "medium", file_name = "density",
  working_dir = paste0(getwd(), "/"), Format = "png", Res = 200,
  add = FALSE, outermargintext = c("Eastings", "Northings"),
  zlim = NULL, col, mar = c(0, 0, 2, 0), oma = c(4, 4, 0, 0),
  legend_x = c(0, 0.05), legend_y = c(0.05, 0.45), cex.legend = 1,
  mfrow, land_color = "grey", n_cells, xlim, ylim, country = NULL,
  contour_nlevels = 0, fun = mean, ...)
}
\arguments{
\item{Y_gt}{matrix where values for every column are plotted as a map}

\item{map_list}{output from \code{FishStatsUtils::make_map_info}}

\item{panel_labels}{vector defining titles to use for each panel; defaults to blank}

\item{projargs}{A character string of projection arguments; the arguments must be entered exactly as in the PROJ.4 documentation; if the projection is unknown, use \code{as.character(NA)}, it may be missing or an empty string of zero length and will then set to the missing value.}

\item{add}{boolean indicating whether to add plot to an existing panel figure, or to define a new panel figure}

\item{outermargintext}{vector defining text to plot in outer margins of panel figure}

\item{zlim}{range for defining bounds of color scale.  If \code{zlim=NULL}, then a constant scale is inferred from the range of \code{Y_gt} and a color-legend is plotted in the last panel.  If \code{zlim=NA} then a different range is used in each panel from the range of \code{Y_gt[,t]} and a color-legend is plotted in every panel.}

\item{legend_x}{two numeric values (generally between 0 and 1, but slightly lower/higher values generate colorbars that are just outside the plotting window) giving left and right-hand location of color legend}

\item{legend_y}{two numeric values (see legend_y) giving bottom and top location of color legend}

\item{land_color}{color for filling in land (use \code{land_color=rgb(0,0,0,alpha=0)} for transparent land)}

\item{country}{a character vector of country names.}

\item{contour_nlevels}{number of levels used when adding contour lines, passed to \code{\link[graphics]{contour}} as argument \code{nlevels}}

\item{fun}{function or character. To determine what values to assign to cells that are covered by multiple spatial features. You can use functions such as \code{min, max}, or \code{mean}, or one of the following character values: \code{'first'}, \code{'last'}, \code{'count'}. The default value is \code{'last'}. In the case of SpatialLines*, \code{'length'} is also allowed (currently for planar coordinate systems only).


If \code{x} represents points, \code{fun} must accept a \code{na.rm} argument, either explicitly or through 'dots'. This means that \code{fun=length} fails, but \code{fun=function(x,...)length(x)} works, although it ignores the \code{na.rm} argument. To use the \code{na.rm} argument you can use a function like this: fun=function(x, na.rm){if (na.rm) length(na.omit(x)) else (length(x)}, or use a function that removes \code{NA} values in all cases, like this function to compute the number of unique values per grid cell "richness": \code{fun=function(x, ...) {length(unique(na.omit(x)))} }. If you want to count the number of points in each grid cell, you can use \code{ fun='count'} or  \code{fun=function(x,...){length(x)}}.

You can also pass multiple functions using a statement like \code{fun=function(x, ...) c(length(x),mean(x))}, in which case the returned object is a RasterBrick (multiple layers).
}

\item{...}{arguments passed to \code{par}}
}
\description{
\code{plot_variable} plots a map and fills in regions with colors to represent intensity in an areal-interpretion of model results

See \url{https://proj.org/operations/projections/index.html} for a list of projections to pass via \code{projargs}. I often prefer \code{projargs='+proj=natearth +lon_0=0 +units=km'} where argument \code{+lon_0} allows the user to center eastings on a specified longitude. If maps are generating visual artefacts, please try using argument \code{country} to simplify the polygons used to represent land features.
}
