% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{PlotLoadings}
\alias{PlotLoadings}
\title{Copy of FishStatsUtils::plot_loadings}
\usage{
PlotLoadings(...)
}
\description{
Included for continuity when using old scripts
}
\details{
Please use \code{?FishStatsUtils::plot_loadings} to see list of arguments and usage
}
