% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rotate_factors.R
\name{rotate_factors}
\alias{rotate_factors}
\title{Rotate results}
\usage{
rotate_factors(
  Cov_jj = NULL,
  L_pj = NULL,
  Psi_sjt = NULL,
  RotationMethod = "PCA",
  testcutoff = 1e-10,
  quiet = FALSE
)
}
\arguments{
\item{Cov_jj}{Covariance calculated from loadings matrix}

\item{L_pj}{Loadings matrix for `p` categories and `j` factors (calculated from \code{Cov_jj} if it is provided)}

\item{Psi_sjt}{Array of factors (1st dimension: spatial knots;  2nd dimension: factors;  3rd dimension:  time)}

\item{RotationMethod}{Method used for rotation when visualing factor decomposition results,
Options: "PCA" (recommended) or "Varimax"}

\item{testcutoff}{tolerance for numerical rounding when confirming that rotation doesn't effect results}
}
\value{
tagged list of outputs
\describe{
  \item{L_pj_rot}{Loadings matrix after rotation}
  \item{Psi_rot}{Factors after rotation}
  \item{Hinv}{Object used for rotation}
  \item{L_pj}{Loadings matrix}
}
}
\description{
\code{rotate_factors} rotates results from a factor model
}
