% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{Rotate_Fn}
\alias{Rotate_Fn}
\title{Copy of FishStatsUtils::Rotate_Fn}
\usage{
Rotate_Fn(...)
}
\description{
Included for continuity when using old scripts
}
\details{
Please use \code{?FishStatsUtils::rotate_factors} to see list of arguments and usage
}
