% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_variable.R
\name{sample_variable}
\alias{sample_variable}
\title{Sample from predictive distribution of a variable}
\usage{
sample_variable(Sdreport, Obj, variable_name, n_samples = 100,
  seed = 123456)
}
\arguments{
\item{Sdreport}{TMB output from `\code{TMB::sdreport(Obj)}`}

\item{Obj}{Fitted TMB object from package `VAST`, i.e., output from `\code{fit_model(...)$tmb_list$Obj}`}

\item{variable_name}{name of variable available in report using \code{Obj$report()}}

\item{n_samples}{number of samples from the joint predictive distribution for fixed and random effects.  Default is 100, which is slow.}

\item{seed}{integer used to set random-number seed when sampling variables, as passed to \code{set.seed(.)}}
}
\description{
\code{sample_variable} samples from the joint distribution of random effects to approximate the predictive distribution for a variable

Exploratory testing suggests that calling \code{Obj$report(newvalues)} within a function like \code{sample_variable(.)} doesn't affect the results of subsequent calls of \code{Obj$report()} so this function does not appear to change anything in the global environment
}
