% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_data.R
\name{simulate_data}
\alias{simulate_data}
\title{Simulate new dynamics and sampling data}
\usage{
simulate_data(fit, type = 1, random_seed = NULL)
}
\arguments{
\item{fit}{output form \code{fit_model(.)}}

\item{type}{integer stating what type of simulation to use.  \code{type=1} simulates new data conditional upon estimated fixed and random effects.  \code{type=2} simulates new random effects conditional upon fixed effects, and new data conditional upon both.  \code{type=3} simulates new fixed and random effects from the joint precision matrix, and new data conditional upon these values.}

\item{random_seed}{integer passed to \code{\link[base]{set.seed}}, where the default value \code{random_seed=NULL} resets the random-number seed.}
}
\value{
Report object containing new data and population variables including
\describe{
  \item{b_i}{New simulated data}
  \item{D_gcy}{Density for each grid cell g, category c, and year y}
  \item{Index_cyl}{Index of abundance for each category c, year y, and stratum l}
}
}
\description{
\code{simulate_data} conducts a parametric bootstrap to simulate new data and potentially simulate new population dynamics and associated variables
}
