% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize_covariance.R
\name{summarize_covariance}
\alias{summarize_covariance}
\title{Explore spatio-temporal covariance}
\usage{
summarize_covariance(Report, Data, ParHat, SD = NULL,
  category_order = 1:Data$n_c, category_names = 1:Data$n_c,
  plotdir = paste0(getwd(), "/"), figname = "Cov", plotTF = NULL,
  plot_cor = TRUE, mgp = c(2, 0.5, 0), tck = -0.02, oma = c(0, 5,
  2, 0), ...)
}
\arguments{
\item{category_order}{integer-vector to re-order categories while plotting}

\item{category_names}{character-vector listing name for each category}

\item{plotdir}{directory (absolute path) for plots}

\item{figname}{character for first-part of figure names}

\item{plotTF}{integer-vector (length 4) specifying which covariances to plot (recommended: use \code{plotTF=FieldConfig})}

\item{plot_cor}{Boolean, whether to plot correlation or covariance}

\item{mgp}{passed to \code{par}}

\item{tck}{passed to \code{par}}

\item{oma}{passed to \code{par}}

\item{...}{passed to \code{ThorsonUtilities::save_fig}}
}
\description{
\code{Summarize_Covariance} plots and returns spatio-temporal covariance among categories
}
