% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Calc_Anisotropic_Mesh.R, R/zzz.R
\name{Calc_Anisotropic_Mesh}
\alias{Calc_Anisotropic_Mesh}
\title{Make mesh for distances among points}
\usage{
Calc_Anisotropic_Mesh(...)

Calc_Anisotropic_Mesh(...)
}
\arguments{
\item{...}{Arguments passed to \code{INLA::inla.mesh.create}}

\item{loc_x}{location (eastings and northings in kilometers, UTM) for each sample or knot}

\item{Method}{spatial method determines ("Mesh" and "Grid" give}

\item{anisotropic_mesh}{OPTIONAL, anisotropic mesh (if missing, its recalculated from loc_x)}
}
\value{
Tagged list containing distance metrics
}
\description{
\code{Calc_Anistropic_Mesh} builds a tagged list representing distances for isotropic or geometric anisotropic triangulated mesh

Included for continuity when using old scripts
}
\details{
Please use \code{?FishStatsUtils::make_mesh} to see list of arguments and usage
}
