% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_proportion.R
\name{calculate_proportion}
\alias{calculate_proportion}
\title{Calculate the compositional-expansion}
\usage{
calculate_proportion(
  TmbData,
  Index,
  Expansion_cz = NULL,
  year_labels = NULL,
  years_to_plot = NULL,
  strata_names = NULL,
  category_names = NULL,
  plot_legend = ifelse(TmbData$n_l > 1, TRUE, FALSE),
  DirName = paste0(getwd(), "/"),
  PlotName = "Proportion.png",
  PlotName2 = "Average.png",
  interval_width = 1,
  width = 6,
  height = 6,
  xlab = "Category",
  ylab = "Proportion",
  ...
)
}
\arguments{
\item{TmbData}{Formatted data inputs, from \code{\link[VAST]{make_data}}}

\item{Index}{output from \code{FishStatsUtils::plot_biomass_index}}

\item{Expansion_cz}{matrix specifying how densities are expanded when calculating annual indices, with a row for each category \code{c} and two columns.
The first column specifies whether to calculate annual index for category \code{c} as the weighted-sum across density estimates,
where density is weighted by area ("area-weighted expansion", \code{Expansion[c,1]=0}, the default),
where density is weighted by the expanded value for another category ("abundance weighted expansion" \code{Expansion[c1,1]=1}),
or the index is calculated as the weighted average of density weighted by the expanded value for another category
("abundance weighted-average expansion" \code{Expansion[c1,1]=2}).  The 2nd column is only used when \code{Expansion[c1,1]=1} or \code{Expansion[c1,1]=2},
and specifies the category to use for abundance-weighted expansion, where \code{Expansion[c1,2]=c2} and \code{c2} must be lower than \code{c1}.}

\item{year_labels}{character vector specifying names for labeling times \code{t_i}}

\item{years_to_plot}{integer vector, specifying positions of \code{year_labels} for plotting (used to avoid plotting years with no data, etc.)}

\item{strata_names}{names for spatial strata}

\item{category_names}{names for categories (if using package \code{`VAST`})}

\item{plot_legend}{Add legend for labelling colors}

\item{DirName}{Directory for saving plot and table}

\item{PlotName}{Name for plot}

\item{interval_width}{width for confidence intervals}

\item{width}{plot width in inches}

\item{height}{plot height in inches}

\item{...}{list of arguments to pass to \code{plot_index}}
}
\value{
Tagged list of output
\describe{
  \item{Prop_ctl}{Proportion of biomass for each category c, time t, and stratum l}
  \item{Neff_tl}{Effective sample size (median across categories) for each time t and stratum l}
}
}
\description{
\code{calculate_proportion} takes output from a VAST run and calculates the proportion of biomass in different categories
}
\references{
For details regarding multivariate index standardization and expansion see \url{https://cdnsciencepub.com/doi/full/10.1139/cjfas-2018-0015}
}
