% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_model.R
\name{fit_model}
\alias{fit_model}
\title{Fit VAST to data}
\usage{
fit_model(settings, Lat_i, Lon_i, t_iz, c_iz, b_i, a_i, v_i,
  working_dir = paste0(getwd(), "/"), Xconfig_zcp = NULL,
  X_gtp = NULL, X_itp = NULL, Q_ik = NULL, newtonsteps = 1,
  extrapolation_args = list(), optimize_args = list(), ...)
}
\arguments{
\item{Lat_i}{Latitude for each sample}

\item{Lon_i}{Longitude for each sample}

\item{t_iz}{Matrix where each row species the time for each observation i (if t_iz is a vector, it is coerced to a matrix with one column; if it is a matrix with two or more columns, it specifies multiple times for each observation, e.g., both year and season)}

\item{c_iz}{Category (e.g., species, length-bin) for each observation i}

\item{b_i}{Sampled biomass for each observation i}

\item{a_i}{Sampled area for each observation i}

\item{v_i}{OPTIONAL, sampling category (e.g., vessel or tow) associated with overdispersed variation for each observation i}

\item{Xconfig_zcp}{OPTIONAL, 3D array of settings for each dynamic density covariate, where the first dimension corresponds to 1st or 2nd linear predictors, second dimension corresponds to model category, and third dimension corresponds to each density covariate
\describe{
  \item{Xconfig_zcp[z,c,p]=0}{\code{X_xtp[,,p]} has no effect on linear predictor z for category c}
  \item{Xconfig_zcp[z,c,p]=1}{\code{X_xtp[,,p]} has a linear effect on linear predictor z for category c}
  \item{Xconfig_zcp[z,c,p]=2}{\code{X_xtp[,,p]} has a spatially varying, zero-centered linear effect on linear predictor z for category c}
  \item{Xconfig_zcp[z,c,p]=3}{\code{X_xtp[,,p]} has a spatially varying linear effect on linear predictor z for category c}
}}

\item{Q_ik}{OPTIONAL, matrix of catchability covariates (e.g., measured variables affecting catch rates but not caused by variation in species density) for each observation i}

\item{newtonsteps}{number of extra newton steps to take after optimization (alternative to \code{loopnum})}

\item{extrapolation_args}{tagged list of optional arguments to pass to \code{make_extrapolation_info}}

\item{optimize_args}{tagged list of optional arguments to pass to \code{TMBhelper::Optimize}}

\item{...}{additional parameters to pass to \code{VAST::make_data}}
}
\description{
\code{fit_model} fits a spatio-temporal model to data
}
