% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_example.R
\name{load_example}
\alias{load_example}
\title{Load an example data set}
\usage{
load_example(data_set = "EBS_pollock")
}
\arguments{
\item{data_set}{data set to load}
}
\description{
\code{load_example} loads a catch and effort data set from multiple surveys worldwide
}
