% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_example.R
\name{load_example}
\alias{load_example}
\title{Load an example data set}
\usage{
load_example(data_set = "EBS_pollock")
}
\arguments{
\item{data_set}{data set to load}
}
\description{
\code{load_example} loads a catch and effort data set from multiple surveys worldwide
}
\details{
Current options for \code{data_set} include "Chatham_rise_hake", "Iceland_cod", "WCGBTS_canary", "GSL_american_plaice", "BC_pacific_cod", "EBS_pollock", "GOA_Pcod", "GOA_pollock", "GB_spring_haddock", "GB_fall_haddock", "SAWC_jacopever", "Aleutian_islands_POP", "condition_and_density", "multimodal_red_snapper", "lingcod_comp_expansion", "covariate_example", and "ordination".  These examples are used to highlight different functionality for spatio-temporal packages, as well as during automated testing of backwards compatibility.
}
