% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_covariates.R
\name{make_covariates}
\alias{make_covariates}
\title{Format habitat covariate matrix}
\usage{
make_covariates(formula = ~0, covariate_data, Year_i, spatial_list,
  extrapolation_list)
}
\arguments{
\item{formula}{an object of class "formula" (or one that can be coerced to that class): a symbolic description of the model to be fitted. Similar specification to \code{\link{stats::lm}}}

\item{covariate_data}{data frame of covariate values with columns \code{Lat}, \code{Lon}, and \code{Year}, and other columns matching names in \code{formula}}
}
\value{
Tagged list of useful output
\describe{
  \item{Cov_gtp}{3-dimensional array for use in \code{VAST::make_data}}
  \item{Cov_itp}{3-dimensional array for use in \code{VAST::make_data}}
}
}
\description{
\code{make_covariates} uses a formula interface to generate covariates
}
\details{
This function generates 3D arrays required by \code{VAST::make_data} to incorporate density covariates. The user must supply a data frame of covariate values \code{covariate_data}, with Lat, Lon, and Year columns, and then covariates at each observation \code{i} are matched against the nearest supplied covariate value in that year. Similarly, covariate values at each extrapolation-grid cell in each year, \code{X_gtp}, are matched against the nearest covariate value in that year.
}
