% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_settings.R
\name{make_settings}
\alias{make_settings}
\title{Make list of settings}
\usage{
make_settings(n_x, Region, purpose = "index", fine_scale = TRUE,
  strata.limits = data.frame(STRATA = "All_areas"), zone = NA)
}
\arguments{
\item{n_x}{the number of vertices in the SPDE mesh (determines the spatial resolution when Method="Mesh")}

\item{Region}{a character entry that is matched against potential values to determine the region for the extrapolation grid.}

\item{strata.limits}{an input for determining stratification of indices (see example script)}
}
\description{
\code{make_settings} makes a list of settings for a given purpose
}
