% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_settings.R
\name{make_settings}
\alias{make_settings}
\title{Make list of settings}
\usage{
make_settings(
  n_x,
  Region,
  purpose = "index",
  fine_scale = TRUE,
  strata.limits = data.frame(STRATA = "All_areas"),
  zone = NA,
  FieldConfig,
  RhoConfig,
  OverdispersionConfig,
  ObsModel,
  bias.correct,
  Options,
  use_anisotropy,
  vars_to_correct,
  Version,
  treat_nonencounter_as_zero,
  n_categories,
  VamConfig,
  max_cells,
  knot_method
)
}
\arguments{
\item{n_x}{the number of vertices in the SPDE mesh (determines the spatial resolution when \code{Method="Mesh"})}

\item{Region}{a character vector, where each element that is matched against potential values to determine the region for the extrapolation grid. Current options are \code{"california_current", "west_coast_hook_and_line", "british_columbia", "eastern_bering_sea", "northern_bering_sea", "bering_sea_slope", "st_matthews_island", "aleutian_islands", "gulf_of_alaska", "northwest_atlantic", "south_africa", "gulf_of_st_lawrence", "new_zealand", "habcam", "gulf_of_mexico", "ATL-IBTS-Q1", "ATL-IBTS-Q4", "BITS", "BTS", "BTS-VIIA", "EVHOE", "IE-IGFS", "NIGFS", "NS_IBTS", "PT-IBTS", "SP-ARSA", "SP-NORTH", "SP-PORC", "stream_network", "user", "other"}, or the absolute path and file name for a GIS shapefile}

\item{purpose}{character indicating what purpose is intended for the model, and therefore what default settings are perhaps appropriate. Only currently implemented for \code{purpose="index"}, \code{purpose="condition_and_density"}, \code{purpose="MICE"}, or \code{purpose="ordination"}.}

\item{fine_scale}{a Boolean indicating whether to ignore (\code{fine_scale=FALSE}) or account for (\code{fine_scale=TRUE}) fine-scale spatial heterogeneity;  See details for more informatino}

\item{strata.limits}{an input for determining stratification of indices (see example script)}

\item{zone}{UTM zone used for projecting Lat-Lon to km distances; use \code{zone=NA} by default to automatically detect UTM zone from the location of extrapolation-grid samples}

\item{FieldConfig}{See Details section of \code{\link[VAST]{make_data}} for details}

\item{RhoConfig}{vector of form \code{c("Beta1"=0,"Beta2"=0,"Epsilon1"=0,"Epsilon2"=0)} specifying whether either intercepts (Beta1 and Beta2)
or spatio-temporal variation (Epsilon1 and Epsilon2) is structured among time intervals (0: each year as fixed effect;
1: each year as random following IID distribution; 2: each year as random following a random walk;
3: constant among years as fixed effect; 4: each year as random following AR1 process);  If missing, assumed to be zero for each element}

\item{OverdispersionConfig}{a vector of format \code{c("eta1"=0, "eta2"="AR1")} governing any correlated overdispersion
among categories for each level of \code{v_i}, where eta1 is for encounter probability, and eta2 is for positive catch rates,
where \code{0} is off, code{"AR1"} is an AR1 process, and aninteger greater than zero (e.g., \code{2}) is the number of elements in a factor-analysis covariance (by default,
\code{c("eta1"=0, "eta2"=0)} and this turns off overdispersion)}

\item{Options}{a tagged-vector that is empty by default, \code{Options=c()}, but where the following slots might be helpful to add,
       either by passing \code{Options} to \code{\link[FishStatsUtils]{make_settings}}, or editing after a call to that function:
\describe{
  \item{\code{Options["SD_site_logdensity"]=TRUE}}{Turns on standard error calculation for local log-density (which is very slow to calculate!)}
  \item{\code{Options["Calculate_Range"]=TRUE}}{Turns on internal calculation and SE for center-of-gravity}
  \item{\code{Options["Calculate_effective_area"]=TRUE}}{Turns on internal calculation and SE for effective area occupied measuring range expansion/contraction}
  \item{\code{Options["Calculate_Cov_SE"]=TRUE}}{Turns on internal calculation and SE for covariance among categories (i.e. in factor model)}
  \item{\code{Options["Calculate_proportion"]=TRUE}}{Turns on internal calculation and SE for proportion of response within each category (e.g., for calculating proportion-at-age or species turnover)}
  \item{\code{Options["Calculate_Synchrony"]=TRUE}}{Turns on internal calculation and SE for Loreau metric of synchrony (a.k.a. portfolio effects)}
}}

\item{use_anisotropy}{Boolean indicating whether to estimate two additional parameters representing geometric anisotropy}

\item{vars_to_correct}{a character-vector listing which parameters to include for bias-correction, as passed to \code{TMBhelper::Optimize}}

\item{Version}{Which CPP version to use.  If missing, defaults to latest version using \code{FishStatsUtils::get_latest_version(package="VAST")}.
Can be used to specify using an older CPP, to maintain backwards compatibility.}

\item{treat_nonencounter_as_zero}{Boolean indicating whether to treat any year-category combination as having zero biomass when generating abundance indices and resulting compositional estimates}

\item{n_categories}{number of categories in a multivariate model (only necessary to specify given some values for \code{purpose})}

\item{VamConfig}{Options to estimate interactions, containing three slots:
\describe{
  \item{\code{VamConfig[0]}}{selects method for forming interaction matrix; Turn off feature using 0, or I recommend using 2 by default}
  \item{\code{VamConfig[1]}}{indicates the rank of the interaction matrix, indicating the number of community axes that have regulated dynamics}
  \item{\code{VamConfig[2]}}{Indicates whether interactions occur before spatio-temporal variation (\code{VamConfig[2]=0}) or after \code{VamConfig[2]=1}}
}}

\item{max_cells}{Maximum number of extrapolation-grid cells.  If number of cells in extrapolation-grid is less than this number, then its value is ignored.  Default \code{max_cells=Inf} results in no reduction in number of grid cells from the default extrapolation-grid for a given region.  Using a lower value is particularly useful when \code{fine_scale=TRUE} and using epsilon bias-correction, such that the number of extrapolation-grid cells is often a limiting factor in estimation speed.}

\item{knot_method}{whether to determine location of GMRF vertices based on the location of samples \code{knot_method=`samples`} or extrapolation-grid cells within the specified strata \code{knot_method='grid'}; default \code{knot_method=NULL} is coerced to \code{knot_method=`samples`}}
}
\value{
Tagged list containing default settings for a given purpose, use \code{names} on output to see list of settings.
}
\description{
\code{make_settings} makes a list of settings for a given purpose
}
\details{
This function assembles a default set of user-decisions for a specified modelling purpose. The default settings are guessed based on generic guidance, and should be carefully reviewed for real-world purposes. If the user supplies values for individual settings e.g. \code{FieldConfig}, then these values override the defaults that are provided by interpreting \code{purpose}
}
\seealso{
\code{\link[VAST]{VAST}} for general documentation, \code{\link[FishStatsUtils]{make_settings}} for generic settings, \code{\link[FishStatsUtils]{fit_model}} for model fitting, and \code{\link[FishStatsUtils]{plot_results}} for generic plots

Other wrapper functions: 
\code{\link{fit_model}()},
\code{\link{plot_results}()}
}
\concept{wrapper functions}
