% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_factors.R
\name{plot_factors}
\alias{plot_factors}
\title{Plot factor-decomposition of covariance}
\usage{
plot_factors(
  fit,
  Report = fit$Report,
  ParHat = fit$ParHat,
  Data = fit$data_list,
  Obj = fit$tmb_list$Obj,
  SD = fit$parameter_estimates$SD,
  year_labels = fit$year_labels,
  category_names = NULL,
  RotationMethod = "PCA",
  mapdetails_list = NULL,
  Dim_year = NULL,
  Dim_species = NULL,
  projargs = fit$extrapolation_list$projargs,
  plotdir = paste0(getwd(), "/"),
  land_color = "grey",
  zlim = NA,
  testcutoff = 1e-04,
  ...
)
}
\arguments{
\item{Report}{output report, e.g., from \code{Report <- obj$report()}}

\item{ParHat}{Tagged list of fitted data, e.g., from \code{ParHat <- obj$env$parList()}}

\item{Data}{tagged list of input data}

\item{Obj}{Fitted TMB object from package `VAST`, i.e., output from `\code{fit_model(...)$tmb_list$Obj}`}

\item{SD}{standard deviation report, e.g., from \code{SD <- sdreport( obj )}}

\item{year_labels}{Year names for labeling panels}

\item{category_names}{character-vector listing name for each category}

\item{RotationMethod}{Method used for rotation, Options: "PCA" (recommended) or "Varimax"}

\item{mapdetails_list}{output from \code{\link{make_map_info}}}

\item{Dim_year}{Plotting dimension (row,column) for plot of years (default: square with sufficient size for number of years)}

\item{Dim_species}{Plotting dimension (row,column) for plot of categories (default: square with sufficient size for number of categories)}

\item{projargs}{a CRS, e.g. "+proj=utm +datum=WGS84 +units=km +zone=3", which is recommended for displaying the entirety of Alaska in a single plot.See \url{https://proj.org/operations/projections/index.html} for a list of projections to pass via \code{projargs}. I often prefer \code{projargs='+proj=natearth +lat_0=0 +units=km'} where argument \code{+lat_0} allows the user to center eastings on a specified latitude.}

\item{plotdir}{directory for saving plots}

\item{land_color}{color for filling in land (use \code{land_color=rgb(0,0,0,alpha=0)} for transparent land)}

\item{zlim}{range for defining bounds of color scale.  If \code{zlim=NULL}, then a constant scale is inferred from the range of the plotted variable and a color-legend is plotted in the last panel.  If \code{zlim=NA} then a different range is used in each panel from the range of \code{Y_gt[,t]} and a color-legend is plotted in every panel.}

\item{testcutoff}{tolerance for numerical rounding when confirming that rotation doesn't effect results}

\item{...}{additional arguments passed to \code{\link{plot_maps}} and/or \code{\link{plot_variable}} when plotting factor-values on a map}
}
\description{
\code{plot_factors} plots factor loadings, average spatial factors, and spatio-temporal factors
}
