% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_quantile_residuals.R
\name{plot_quantile_residuals}
\alias{plot_quantile_residuals}
\title{Plot quantile residuals on map}
\usage{
plot_quantile_residuals(
  dharmaRes,
  fit,
  file_name = "quantile_residuals_on_map",
  zlim = NULL,
  year_labels = NULL,
  years_to_plot = NULL,
  n_cells = NULL,
  ...
)
}
\arguments{
\item{zlim}{range for defining bounds of color scale.  If \code{zlim=NULL}, then a constant scale is inferred from the range of the plotted variable and a color-legend is plotted in the last panel.  If \code{zlim=NA} then a different range is used in each panel from the range of \code{Y_gt[,t]} and a color-legend is plotted in every panel.}

\item{year_labels}{Year names for labeling panels}

\item{years_to_plot}{integer vector, specifying positions of \code{year_labels} for plotting (used to avoid plotting years with no data, etc.)}

\item{n_cells}{Integer used to determine the argument \code{cell.size} passed
to \code{\link[plotKML]{vect2rast}} when converting output at extrapolation-grid cells
to a raster prior to plotting mapped outputs using \code{\link{plot_variable}}.  An increased \code{n_cells} results in a decreased
\code{cell.size} and hence higher resolution plots.  The default value for \code{n_cells}
will often need to be modified for a given purpose.}

\item{...}{arguments passed to \code{\link{plot_variable}}}

\item{output}{from \code{summary.fit_model(x,what="residuals")}}

\item{x}{Output from \code{\link{fit_model}}}
}
\description{
\code{plot_quantile_residuals} shows aggregated quantile residuals at a user-specified spatial resolution

This function plots quantile residuals on a map, and is useful to check for persistent spatial patterns
in those residuals.  For a check for outliers (i.e., the scale of residuals) please see the separate Q-Q plot.
See \code{\link{summary.fit_model}} for more details regarding inputs
}
