% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_quantile_residuals.R
\name{plot_quantile_residuals}
\alias{plot_quantile_residuals}
\title{Plot quantile residuals on map}
\usage{
plot_quantile_residuals(dharmaRes, fit,
  file_name = "quantile_residuals_on_map", Year_Set = NULL,
  Years2Include = NULL, ...)
}
\arguments{
\item{...}{arguments passed to \code{FisshStatsUtils::plot_variable}}

\item{x}{Output from \code{\link{fit_model}}}
}
\description{
\code{plot_quantile_residuals} shows aggregated quantile residuals at a user-specified spatial resolution
}
