% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_results.R
\name{plot_results}
\alias{plot_results}
\title{Plot results}
\usage{
plot_results(fit, settings, working_dir = paste0(getwd(), "/"),
  year_labels = NULL, years_to_plot = NULL)
}
\description{
\code{plot_results} plots diagnostics, results, and indices for a given fitted model
}
