% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_results.R
\name{plot_results}
\alias{plot_results}
\title{Plot results}
\usage{
plot_results(fit, settings = fit$settings, plot_set = 3,
  working_dir = paste0(getwd(), "/"), year_labels = fit$year_labels,
  years_to_plot = fit$years_to_plot, use_biascorr = TRUE, map_list,
  check_residuals = TRUE, projargs = "+proj=longlat", ...)
}
\arguments{
\item{fit}{Output from \code{fit_model}}

\item{settings}{Output from \code{make_settings}}

\item{plot_set}{integer-vector defining plots to create
\describe{
  \item{plot_set=1}{Probability of encounter/non-encounter}
  \item{plot_set=2}{Log-expected positive catch rate}
  \item{plot_set=3}{Log-predicted density (product of encounter probability and positive catch rates)}
  \item{plot_set=4}{Log-positive catch rates (rescaled)}
  \item{plot_set=5}{Log-predicted density (rescaled)}
  \item{plot_set=6}{Spatio-temporal variation in encounter probability}
  \item{plot_set=7}{Spatio-temporal variation in log-positive catch rates}
  \item{plot_set=8}{Linear predictor for encounter probability}
  \item{plot_set=9}{Linear predictor for positive catch rates}
  \item{plot_set=10}{Coefficient of variation for predicted density (available only if \code{Data_Fn(...,Options=c('SD_site_logdensity'=1,...))}}
  \item{plot_set=11}{Covariates that are included in the model}
  \item{plot_set=12}{Total biomass across all categories (only useful in a multivariate model)}
  \item{plot_set=13}{Covariate effects on encounter probability}
  \item{plot_set=14}{Covariate effects on positive catch rates}
}}

\item{check_residuals}{Boolean indicating whether to run or skip residual diagnostic plots (which can be slow as currently implemented)}

\item{...}{additional settings to pass to \code{FishStatsUtils::plot_maps}}
}
\value{
Invisibly returns a tagged list of outputs generated by standard plots.
}
\description{
\code{plot_results} plots diagnostics, results, and indices for a given fitted model
}
\details{
This function takes a fitted VAST model and generates a standard set of diagnostic and visualization plots.
}
\seealso{
\code{?VAST} for general documentation, \code{?make_settings} for generic settings, \code{?fit_model} for model fitting, and \code{?plot_results} for generic plots

Other wrapper functions: \code{\link{fit_model}},
  \code{\link{make_settings}}
}
\concept{wrapper functions}
