% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_variable.R
\name{plot_variable}
\alias{plot_variable}
\title{Plot results on a multi-panel figure of maps}
\usage{
plot_variable(Y_gt, map_list, panel_labels, projargs = "+proj=longlat",
  map_resolution = "medium", file_name = "density",
  working_dir = paste0(getwd(), "/"), Format = "png", Res = 200,
  add = FALSE, outermargintext = c("Eastings", "Northings"), zlim, col,
  mar = c(0, 0, 2, 0), oma = c(4, 4, 0, 0), legend_x = c(0, 0.05),
  legend_y = c(0.05, 0.45), mfrow, land_color = "grey", ...)
}
\arguments{
\item{Y_gt}{matrix where values for every column are plotted as a map}

\item{map_list}{output from \code{FishStatsUtils::make_map_info}}

\item{panel_labels}{vector defining titles to use for each panel; defaults to blank}

\item{projargs}{A character string of projection arguments; the arguments must be entered exactly as in the PROJ.4 documentation; if the projection is unknown, use \code{as.character(NA)}, it may be missing or an empty string of zero length and will then set to the missing value.}

\item{add}{boolean indicating whether to add plot to an existing panel figure, or to define a new panel figure}

\item{outermargintext}{vector defining text to plot in outer margins of panel figure}

\item{zlim}{range for defining bounds of color scale}

\item{legend_x}{two numeric values (between 0 and 1) giving left and right-hand location of color legend}

\item{legend_y}{two numeric values (between 0 and 1) giving bottom and top location of color legend}

\item{land_color}{color for filling in land (use \code{land_color=rgb(0,0,0,alpha=0)} for transparent land)}

\item{...}{arguments passed to \code{par}}
}
\description{
\code{plot_variable} plots a map and fills in regions with colors to represent intensity in an areal-interpretion of model results
}
