% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rotate_Fn.R
\name{Rotate_Fn}
\alias{Rotate_Fn}
\title{Rotate results}
\usage{
Rotate_Fn(Cov_jj = NULL, L_pj = NULL, Psi, RotationMethod = "PCA",
  testcutoff = 1e-10)
}
\arguments{
\item{Cov_jj}{Covariance calculated from loadings matrix}

\item{L_pj}{Loadings matrix for `p` categories and `j` factors (calculated from \code{Cov_jj} if it is provided)}

\item{Psi}{Array of factors (1st dimension: spatial knots;  2nd dimension: factors;  3rd dimension:  time)}

\item{RotationMethod}{Method used for rotation, Options: "PCA" (recommended) or "Varimax"}

\item{testcutoff}{tolerance for numerical rounding when confirming that rotation doesn't effect results}
}
\value{
tagged list of outputs
\describe{
  \item{L_pj_rot}{Loadings matrix after rotation}
  \item{Psi_rot}{Factors after rotation}
  \item{Hinv}{Object used for rotation}
  \item{L_pj}{Loadings matrix}
}
}
\description{
\code{Rotate_Fn} rotates results from a factor model
}

