% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_model.R
\name{summary.fit_model}
\alias{summary.fit_model}
\title{Extract spatial estimates}
\usage{
\method{summary}{fit_model}(x, what = "density", ...)
}
\arguments{
\item{...}{Not used}

\item{fit}{Output from \code{\link{fit_model}}}
}
\description{
Extract summary of spatial estimates
}
